Data-Science-Capstone-Website-updated/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
asttokens 
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.9.1
colorama 
comm
cryptography==42.0.5
debugpy 
decorator 
exceptiongroup 
executing 
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
idna==3.7
importlib_metadata 
ipykernel
ipython
jedi 
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client 
jupyter_core 
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline 
mdurl==0.1.2
multidict==6.0.5
mysql==0.0.3
mysql-connector
mysqlclient
nest_asyncio 
numpy
packaging 
pandas
parso 
pickleshare 
pillow==10.3.0
platformdirs 
prompt-toolkit 
protobuf==4.25.3
psutil 
pure-eval 
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.0b1
Pygments
PyMySQL==1.1.0
python-dateutil 
python-dotenv==1.0.1
python-tds==1.15.0
pytz
pyzmq 
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six
smmap==5.0.1
SQLAlchemy==2.0.7
sqlalchemy-pytds==0.3.5
stack-data 
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado
traitlets 
typing_extensions
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth
yarl==1.9.4
zipp 
